refactor(Checklist): Checklist
Checklist 폴더를 domain 폴더 하위로 이동하였습니다.
