+++ src/main/java/com/space/server/domin/checklist/service/implementation/ChecklistCreator.java
@@ -1,7 +1,7 @@
-package com.space.server.core.checklist.service.implementation;
+package com.space.server.domin.checklist.service.implementation;
 
-import com.space.server.core.checklist.domain.Checklist;
-import com.space.server.core.checklist.domain.repository.ChecklistRepository;
+import com.space.server.domin.checklist.domain.Checklist;
+import com.space.server.domin.checklist.domain.repository.ChecklistRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
