+++ src/main/java/com/space/server/domin/checklist/service/implementation/ChecklistReader.java
@@ -1,8 +1,8 @@
-package com.space.server.core.checklist.service.implementation;
+package com.space.server.domin.checklist.service.implementation;
 
-import com.space.server.core.checklist.domain.Checklist;
-import com.space.server.core.checklist.domain.repository.ChecklistRepository;
-import com.space.server.core.checklist.exception.ChecklistNotFoundException;
+import com.space.server.domin.checklist.domain.Checklist;
+import com.space.server.domin.checklist.domain.repository.ChecklistRepository;
+import com.space.server.domin.checklist.exception.ChecklistNotFoundException;
 import com.space.server.domin.quiz.domain.Quiz;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
